Refactors redis env vars, so they can be overridden.
* Moves environment and project env vars out of the hook, so they always override.
