+++ web/modules/custom/shepherd/shp_redis_support/shp_redis_support.module
@@ -1,0 +1,30 @@
+<?php
+
+/**
+ * @file
+ * Contains shp_redis_support.module.
+ */
+
+use Drupal\node\NodeInterface;
+
+/**
+ * Implements hook_shp_env_vars().
+ */
+function shp_redis_support_shp_env_vars(NodeInterface $environment) {
+  $env_vars = [];
+  /** @var Drupal\shp_orchestration\OrchestrationProviderInterface $orchestration_provider_plugin */
+  $orchestration_provider = \Drupal::service('plugin.manager.orchestration_provider')
+    ->getProviderInstance();
+  if ($orchestration_provider->getPluginId() === 'openshift_with_redis') {
+    $site = \Drupal::service('shp_custom.environment')->getSite($environment);
+    $project = \Drupal::service('shp_custom.site')->getProject($site);
+    $deployment_name = $orchestration_provider::generateDeploymentName(
+      $project->getTitle(),
+      $site->field_shp_short_name->value,
+      $environment->id()
+    );
+    $env_vars['REDIS_ENABLED'] = '1';
+    $env_vars['REDIS_HOST'] = $deployment_name . '-redis';
+  }
+  return $env_vars;
+}
