+++ web/modules/custom/shepherd/shp_redis_support/src/EventSubscriber/DeploymentEventSubscriber.php
@@ -25,13 +25,7 @@ class DeploymentEventSubscriber implements EventSubscriberInterface {
    * @param \Drupal\shp_orchestration\Event\OrchestrationEnvironmentEvent $event
    */
   public function setupRedisDeployment(OrchestrationEnvironmentEvent $event) {
-    $orchestration_provider = $event->getOrchestrationProvider();
-    if ($orchestration_provider->getPluginId() === 'openshift_with_redis') {
-      $event->setEnvironmentVariables([
-        'REDIS_ENABLED' => '1',
-        'REDIS_HOST'    => $event->getDeploymentName() . '-redis',
-      ]);
-    }
+
   }
 
   /**
